+++ R/components.R
@@ -23,7 +23,7 @@
 #' 
 #' @rdname components
 #' @export
-components.mbl_df <- function(object, ...){
+components.mdl_df <- function(object, ...){
   object <- tidyr::pivot_longer(object, all_of(mable_vars(object)),
                              names_to = ".model", values_to = ".fit")
   kv <- key_vars(object)