+++ R/report.R
@@ -11,7 +11,7 @@ report <- function(object, ...){
 }
 
 #' @export
-report.mbl_df <- function(object, ...){
+report.mdl_df <- function(object, ...){
   if(NROW(object) > 1 || length(mable_vars(object)) > 1){
     warning("Model reporting is only supported for individual models, so a glance will be shown. To see the report for a specific model, use `select()` and `filter()` to identify a single model.")
     return(glance(object))