+++ R/zzz.R
@@ -7,7 +7,7 @@
   register_s3_method("pillar", "pillar_shaft", "agg_vec")
   
   register_s3_method("tibble", "tbl_sum", "dcmp_ts")
-  register_s3_method("tibble", "tbl_sum", "mbl_df")
+  register_s3_method("tibble", "tbl_sum", "mdl_df")
   register_s3_method("tibble", "tbl_sum", "fbl_ts")
   
   register_s3_method("ggplot2", "scale_type", "agg_vec")